Convert the LWRP to a custom resource
