+++ metadata.rb
@@ -5,8 +5,10 @@ license          'all_rights'
 description      'Installs and Configures Supermarket from the Omnibus packages on packages.chef.io'
 source_url       'https://github.com/irvingpop/supermarket-omnibus-cookbook'
 issues_url       'https://github.com/irvingpop/supermarket-omnibus-cookbook/issues'
+chef_version     '>= 12.1'
 version          '1.4.1'
 
-depends 'chef-ingredient', '>= 0.18.0'
+depends 'chef-ingredient', '>= 0.19.0'
+depends 'compat_resource', '>= 12.14'
 depends 'hostsfile'
 depends 'fancy_execute'
